+++ code/src/app/api/registry/[slug]/route.ts
@@ -383,6 +383,9 @@ function makeExternalEndpoint(
     case "tasks": {
       const contractType: string = searchParams.get("type");
       const idOrTimestamp: string = searchParams.get("idOrTimestamp");
+      if (contractType == "task") {
+        return `${agentBaseApi}/contracts/task/${idOrTimestamp}`;
+      }
       const filters: string = encodeFilters(searchParams.get("filters"));
       return `${agentBaseApi}/contracts/service/${idOrTimestamp}?type=${contractType}${filters}`;
     }